requirements.txt
 #------Maine------#
###git+https://github.com/Hydra-sjz/pyrogram.git
Pyrogram==2.0.106
telethon~=1.14.0
tgcrypto
python-dateutil
python-dotenv
requests
asgiref

#------youtube------#
git+https://github.com/yt-dlp/yt-dlp.git
YoutubeTags
youtube_search
youtube-dl
youtube_dl
youtube-search-python

pykeyboard
aiohttp==3.9.2 
asyncio
motor==2.5.1
aiofiles
pymongo[srv]==3.12.3
#--------Pillow>=4.2.0
####Pillow>=4.2.0
pillow==9.5.0
wget
httpx[http2]
pySmartDL
cachetools
pyrate_limiter
psutil
asgiref
cloudscraper
Flask==2.1.1
gunicorn==20.1.0
mutagen
bs4

#====ffmpeg-python==0.2.0====#
ffmpeg==1.4
python-ffmpeg==2.0.4
ffprobe==0.5

python-arq
pyyaml
psutil
pytz
##git+https://github.com/Hydra-sjz/iytdl
termcolor
pysocks
fake-useragent
hachoir
pyquery
urllib3
dnspython
brotli
colorama
lxml
unidecode
wikipedia
RyuzakiLib
uvloop
aria2p
gevent
